# New Q4_0 implementation using 2x F16 instead of 1x F32