@ggerganov 2023-04-17
@sw 
Yes, let's do that. I will reimplement this using `Q4_2` as suggested.

Btw, I'm again reconsidering keeping the SIMD `quantize` / `dequantize`. Too much code for little gains.
In #951 we considered keeping support due to LoRA, but I think we have to measure how much it takes to do LoRA with and without SIMD quantization. It might turn out that the difference is very small compared to overall computation. cc @slaren 

Dropping SIMD quantization support would make changes as in this PR much simpler.