@slaren 2023-04-17
Here is a quick test of the performance impact of quantization on LoRA:
```
perf_total_per_op_us[             ADD] =  12.776 ms
perf_total_per_op_us[         MUL_MAT] =  47.818 ms
perf_total_per_op_us[           SCALE] =   9.319 ms
perf_total_per_op_us[             CPY] =  45.780 ms
```

The quantization happens on the CPY operation, and currently this represents about about 40% of the time to apply a layer.

Changing `ggml_cpy` to use `quantize_row_q_reference` instead:
```
perf_total_per_op_us[             ADD] =  13.698 ms
perf_total_per_op_us[         MUL_MAT] =  46.508 ms
perf_total_per_op_us[           SCALE] =   9.733 ms
perf_total_per_op_us[             CPY] = 384.500 ms
```
Without SIMD, the CPY takes 80% of the time.

Caveat: `ggml_cpy` is still not multithreaded, so we could expect significant gains by doing that. However, I think that the difference is still going to be significant.