@ggerganov 2023-04-17
If we change the `roundf` to:

```
            const uint8_t vi0 = (uint8_t)(v0 + 8.5f);
            const uint8_t vi1 = (uint8_t)(v1 + 8.5f);
```

The timing should be much closer I think.
But in any case, 0.5s for applying a LoRA adapter does not sound bad at all.

We can reduce `ggml.c` significantly with a few macros. For example:

```c
#define GGML_TENSOR_SIZES_SRC0(x) // defines ne00, ne01, ..., nb00, nb01, etc..
#define GGML_TENSOR_SIZES_SRC1(x) // defines ne10, ne11, ..., nb10, nb11, etc..
#define GGML_TENSOR_SIZES_DST(x)  // defines ne0, ne1, ..., nb0, nb1, etc..

#define GGML_GET_PTR_ROW(x, i) // get ptr to ith row using strides

etc..
```

We can do this refactoring soon, let's say after the quantization work is done.
It should make `ggml.c` much more compact.

But, the SIMD remains a problem because even if you had multiple small files or `ggml.c` wasn't so big, you still need to implement all the missing routines for each new experiment.