@slaren 2023-04-17
>But in any case, 0.5s for applying a LoRA adapter does not sound bad at all.

This is just for a single tensor, applying it to the entire model can take a lot of time, especially with the larger models. Applying the entire LoRA (this is [baize-lora-7B](https://huggingface.co/project-baize/baize-lora-7B)) takes 30 seconds on my machine. This is a bit of a pathological case since it also modifies the feed-forward tensors (usually LoRAs only modify the attention tensors), but still, this is very slow for something that has to be done every time.

>But, the SIMD remains a problem because even if you had multiple small files or ggml.c wasn't so big, you still need to implement all the missing routines for each new experiment.

We can consider the SIMD implementations of functions like quantize and dequantize very low priority and just plainly ignore them in experiments, and only if the experiment is successful, we can allow other people to implement them later on in other PRs. I think this is what we are already doing in practice.