@howard0su 2023-04-18
> @sw Yes, let's do that. I will reimplement this using `Q4_2` as suggested.
> 
> Btw, I'm again reconsidering keeping the SIMD `quantize` / `dequantize`. Too much code for little gains. In #951 we considered keeping support due to LoRA, but I think we have to measure how much it takes to do LoRA with and without SIMD quantization. It might turn out that the difference is very small compared to overall computation. Also, integration of RMSE-optimized quantization similar to #896 would be much simpler. cc @slaren
> 
> Dropping SIMD quantization support would make changes as in this PR much simpler.

One thing to consider is openmp which supports multi-thread and SIMD today. It will keep the code simple, portable while leveraging latest hardware features (multi core and AVX).